Split logic, networks into nnetworks.py, dqn logic stays in dqn.py
